Let GameInterfaceRunner let others know what's up
After a few strolls back-and-forth, I came to the conclusion to
event-drive everything that the GameInterfaceRunner does. This gives us
a clear interface to implement, which in turn makes it a lot easier to
build sound handling, graphics handling or maybe some statistics
collector that tracks how many times something happens in the game.