Can now move top piece of tower and see that it returns upon releasing
Introduced a GameRunner which connects the input listeners with the game
interface. Not sure if this is a good idea moving forward, but it solved
the problem for now. We should get back to this and make sure the
over-all structure is clean though.

Also implemented the missing InputManager as ScreenInputManager.

A nice thing about the listeners is that it was easy to attach an
"EventLogger" to see what happens in the game. Seems to work nicely!